Use attribute to define runner

+++ lib/with_model.rb
@@ -6,16 +6,25 @@ require 'with_model/table'
 require 'with_model/version'
 
 module WithModel
+  class << self
+    attr_writer :runner
+  end
+
+  def self.runner
+    @runner ||= :rspec
+  end
+
   # @param [Symbol] name The constant name to assign the model class to.
   # @param scope Passed to `before`/`after` in the test context. RSpec only.
   # @param options Passed to {WithModel::Model#initialize}.
   # @param block Yielded an instance of {WithModel::Model::DSL}.
   def with_model(name, scope: nil, **options, &block)
+    runner = options.delete(:runner)
     model = Model.new name, **options
     dsl = Model::DSL.new model
     dsl.instance_exec(&block) if block
 
-    setup_object(model, scope)
+    setup_object(model, scope: scope, runner: runner)
   end
 
   # @param [Symbol] name The table name to create.
@@ -23,17 +32,19 @@ module WithModel
   # @param options Passed to {WithModel::Table#initialize}.
   # @param block Passed to {WithModel::Table#initialize} (like {WithModel::Model::DSL#table}).
   def with_table(name, scope: nil, **options, &block)
+    runner = options.delete(:runner)
     table = Table.new name, options, &block
 
-    setup_object(table, scope)
+    setup_object(table, scope: scope, runner: runner)
   end
 
   private
 
   # @param [Object] object The new model object instance to create
   # @param scope Passed to `before`/`after` in the test context. Rspec only.
-  def setup_object(object, scope = nil)
-    if Module.const_defined? 'RSpec'
+  def setup_object(object, scope: nil, runner: nil)
+    case runner || WithModel.runner
+    when :rspec
       before(*scope) do
         object.create
       end
@@ -41,12 +52,14 @@ module WithModel
       after(*scope) do
         object.destroy
       end
-    elsif Module.const_defined? 'Minitest'
+    when :minitest
       object.create
 
       Minitest.after_run do
         object.destroy
       end
+    else
+      raise ArgumentError, "Unsupported test runner set, expected :rspec or :minitest"
     end
   end
 end
